requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-dotenv
requests
pytest
